requirements.txt
python-telegram-bot==21.5
python-dotenv>=1.0.0
pytest>=8.0.0
ocrmypdf>=16.0.0
pdf2docx>=0.5.0
groq>=0.11.0
python-docx>=1.1.0
pillow>=10.0.0
pdf2image>=1.17.0